Refactoring code
Removed don`t use methods or entities, also rename variables, and delete almoste all opportunites getting userId from jwt

+++ backend/src/good_deed/good.deed.service.ts
@@ -11,6 +11,9 @@ import { Result } from 'src/entities/result';
 import { GoodDeedUpdateDto } from 'src/entities/good_deed_dto/good.geed.update.dto';
 import { GoodDeedDto } from 'src/entities/good_deed_dto/good.deed.dto';
 import { IUserFriendService } from 'src/interfaces/services/user.friend.service.interfaces';
+import { PaginationParameter } from 'src/entities/pagination/pagination.parameters';
+import { MetaData } from 'src/entities/pagination/meta.data';
+import { GoodDeedsWithPagingMetaData } from 'src/entities/wrapers/good.deeds.with.paging.meta.data';
 
 @Injectable()
 export class GoodDeedService implements IGoodDeedService {
@@ -20,66 +23,51 @@ export class GoodDeedService implements IGoodDeedService {
     @InjectMapper() private readonly mapper: Mapper) {}
 
     private goodDeedList: GoodDeedDto[] = []; 
-    private friendGoodDeedList: string[] = [];
 
-    async createGoodDeed(goodDeedCreate: GoodDeedCreateDto, userId: string): Promise<Result> {
+    async createGoodDeed(goodDeedCreate: GoodDeedCreateDto): Promise<Result> {
         const goodDeed = this.mapper.map(goodDeedCreate, GoodDeedCreateDto, GoodDeed)
-        goodDeed.UserId = userId;
+
 
         const newGoodDeed = await this.goodDeedRepository.create(goodDeed);
 
         if (!newGoodDeed) {
-            return { isSuccess: false, statusCode: HttpStatus.INTERNAL_SERVER_ERROR, message: `Internal server error.` };
+            return { IsSuccess: false, StatusCode: HttpStatus.INTERNAL_SERVER_ERROR, Message: `Internal server error.` };
         }
 
-        return { isSuccess: true, statusCode: HttpStatus.CREATED, message: `Good deed has been created.` };
+        return { IsSuccess: true, StatusCode: HttpStatus.CREATED, Message: `Good deed has been created.` };
     }
 
-    async findByUserId(userId: string): Promise<GoodDeedDto[] | null> {
+    async findByUserId(userId: string,  pagingParam: PaginationParameter): Promise<GoodDeedsWithPagingMetaData | null> {
         const userGoodDeeds = await this.goodDeedRepository.findByUserId(userId);
 
         if (!userGoodDeeds) {
             return null;
         }
 
+        const start = (pagingParam.currentPage - 1) * pagingParam.pageSize ;
+        const end = start + Number(pagingParam.pageSize);
+        const slicedGoodDeedsArray = userGoodDeeds.slice(start, end);
+    
+        const metaData = new MetaData(pagingParam.pageSize, pagingParam.currentPage, userGoodDeeds.length)
+
         this.goodDeedList.splice(0)
 
-        userGoodDeeds.forEach(element => {
+        slicedGoodDeedsArray.forEach(element => {
             this.goodDeedList.push({ Id: element.Id, GoodDeed: element.GoodDeed });
         });
 
-        return this.goodDeedList;
+        return { GoodDeeds: slicedGoodDeedsArray, MetaData: metaData };
     }
 
-    async findFriendGoodDeeds(userId: string, friendId: string): Promise<string[] | null> {
-        if (!await this.userFriendService.isFriendship({ UserId: userId, FriendId: friendId })) {
-            return null;
-        }
-
-        const friendGoodDeed = await this.findByUserId(friendId);
-
-        if (!friendGoodDeed) {
-            return null;
-        }
-
-        this.friendGoodDeedList.splice(0)
-
-        friendGoodDeed.forEach(element => {
-            this.friendGoodDeedList.push(element.GoodDeed);
-        });
-
-        return this.friendGoodDeedList;
-    }
-
-    async updateGoodDeed(goodDeedUpdate: GoodDeedUpdateDto, userId: string): Promise<Result> {
+    async updateGoodDeed(goodDeedUpdate: GoodDeedUpdateDto): Promise<Result> {
         const foundGoodDeed = await this.goodDeedRepository.findById(goodDeedUpdate.Id);
         
         if (!foundGoodDeed) {
-            return { isSuccess: false, statusCode: HttpStatus.NOT_FOUND, message: `Good deed is not exist.` };
+            return { IsSuccess: false, StatusCode: HttpStatus.NOT_FOUND, Message: `Good deed is not exist.` };
         }
 
-        if (foundGoodDeed.UserId != userId) {
-            return { isSuccess: false, statusCode: HttpStatus.BAD_REQUEST, message: `This user don\`t have such good deed.` };   
+        if (foundGoodDeed.UserId != goodDeedUpdate.UserId) {
+            return { IsSuccess: false, StatusCode: HttpStatus.BAD_REQUEST, Message: `This user don\`t have such good deed.` };   
         }
 
         const updateGoodDeed = this.mapper.map(goodDeedUpdate, GoodDeedUpdateDto, GoodDeed);
@@ -87,25 +75,29 @@ export class GoodDeedService implements IGoodDeedService {
         const countUpdatedGoodDeed = await this.goodDeedRepository.update(updateGoodDeed);
 
         if (countUpdatedGoodDeed != 1) {
-            return { isSuccess: false, statusCode: HttpStatus.INTERNAL_SERVER_ERROR, message: `Internal server error.` };
+            return { IsSuccess: false, StatusCode: HttpStatus.INTERNAL_SERVER_ERROR, Message: `Internal server error.` };
           }
       
-          return { isSuccess: true, statusCode: HttpStatus.CREATED, message: `Good deed has been updated.` };
+          return { IsSuccess: true, StatusCode: HttpStatus.CREATED, Message: `Good deed has been updated.` };
     }
 
-    async deleteGoodDeed(goodDeedDelete: GoodDeedDeleteDto, userId: string): Promise<Result> {
-        const foundGoodDeed = await this.goodDeedRepository.findByUserIdAndDeed(userId, goodDeedDelete.GoodDeed);
+    async deleteGoodDeed(goodDeedDelete: GoodDeedDeleteDto): Promise<Result> {
+        const foundGoodDeed = await this.goodDeedRepository.findById(goodDeedDelete.Id);
 
         if (!foundGoodDeed) {
-          return { isSuccess: false, statusCode: HttpStatus.NOT_FOUND, message: `Good deed don\`t exist.` };
+          return { IsSuccess: false, StatusCode: HttpStatus.NOT_FOUND, Message: `Good deed don\`t exist.` };
+        }
+
+        if (foundGoodDeed.UserId !== goodDeedDelete.UserId) {
+            return { IsSuccess: false, StatusCode: HttpStatus.BAD_REQUEST, Message: `This user don\`t have such good deed.` };   
         }
     
         const deletedGoodDeed = await this.goodDeedRepository.delete(foundGoodDeed);
     
         if (!deletedGoodDeed) {
-          return { isSuccess: false, statusCode: HttpStatus.INTERNAL_SERVER_ERROR, message: `Internal server error.` };
+          return { IsSuccess: false, StatusCode: HttpStatus.INTERNAL_SERVER_ERROR, Message: `Internal server error.` };
         }
     
-        return { isSuccess: true, statusCode: HttpStatus.OK, message: `Good deed has been deleted.` };
+        return { IsSuccess: true, StatusCode: HttpStatus.OK, Message: `Good deed has been deleted.` };
     }
 }
